# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/inventory/categories": {
      "post": {
        "operationId": "InventoryController_createCategory",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Registrar categoría de recurso",
        "tags": [
          "Inventario"
        ],
        "x-implementation-status": "implemented"
      },
      "get": {
        "operationId": "InventoryController_listCategories",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Consultar categorías de recurso",
        "tags": [
          "Inventario"
        ],
        "x-implementation-status": "implemented"
      }
    },
    "/inventory/categories/{categoryId}": {
      "get": {
        "operationId": "InventoryController_getCategory",
        "parameters": [
          {
            "name": "categoryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Consultar categoría de recurso",
        "tags": [
          "Inventario"
        ],
        "x-implementation-status": "implemented"
      },
      "patch": {
        "operationId": "InventoryController_updateCategory",
        "parameters": [
          {
            "name": "categoryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Modificar categoría de recurso",
        "tags": [
          "Inventario"
        ],
        "x-implementation-status": "implemented"
      },
      "delete": {
        "operationId": "InventoryController_deleteCategory",
        "parameters": [
          {
            "name": "categoryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Eliminar categoría de recurso",
        "tags": [
          "Inventario"
        ],
        "x-implementation-status": "implemented"
      }
    },
    "/inventory/resources": {
      "post": {
        "operationId": "InventoryController_createResource",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateResourceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Registrar recurso",
        "tags": [
          "Inventario"
        ],
        "x-implementation-status": "implemented"
      },
      "get": {
        "operationId": "InventoryController_listResources",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "available",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Consultar recursos",
        "tags": [
          "Inventario"
        ],
        "x-implementation-status": "implemented"
      }
    },
    "/inventory/resources/{resourceId}": {
      "get": {
        "operationId": "InventoryController_getResource",
        "parameters": [
          {
            "name": "resourceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Consultar recurso",
        "tags": [
          "Inventario"
        ],
        "x-implementation-status": "implemented"
      },
      "patch": {
        "operationId": "InventoryController_updateResource",
        "parameters": [
          {
            "name": "resourceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateResourceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Modificar recurso",
        "tags": [
          "Inventario"
        ],
        "x-implementation-status": "implemented"
      },
      "delete": {
        "operationId": "InventoryController_deleteResource",
        "parameters": [
          {
            "name": "resourceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Eliminar recurso",
        "tags": [
          "Inventario"
        ],
        "x-implementation-status": "implemented"
      }
    },
    "/inventory/resources/{resourceId}/copies": {
      "post": {
        "operationId": "InventoryController_createCopy",
        "parameters": [
          {
            "name": "resourceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCopyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Registrar ejemplar",
        "tags": [
          "Inventario"
        ],
        "x-implementation-status": "implemented"
      },
      "get": {
        "operationId": "InventoryController_listCopiesByResource",
        "parameters": [
          {
            "name": "resourceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Consultar ejemplares de un recurso",
        "tags": [
          "Inventario"
        ],
        "x-implementation-status": "implemented"
      }
    },
    "/inventory/copies/{copyId}": {
      "get": {
        "operationId": "InventoryController_getCopy",
        "parameters": [
          {
            "name": "copyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Consultar ejemplar",
        "tags": [
          "Inventario"
        ],
        "x-implementation-status": "implemented"
      }
    },
    "/inventory/copies/{copyId}/state": {
      "patch": {
        "operationId": "InventoryController_updateCopyState",
        "parameters": [
          {
            "name": "copyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCopyStateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Modificar estado de un ejemplar",
        "tags": [
          "Inventario"
        ],
        "x-implementation-status": "implemented"
      }
    },
    "/inventory/resources/{resourceId}/availability": {
      "get": {
        "operationId": "InventoryController_getResourceAvailability",
        "parameters": [
          {
            "name": "resourceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Consultar disponibilidad de un recurso",
        "tags": [
          "Inventario"
        ],
        "x-implementation-status": "implemented"
      }
    },
    "/inventory/copies/{copyId}/observations": {
      "post": {
        "operationId": "InventoryController_createObservation",
        "parameters": [
          {
            "name": "copyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCopyObservationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Registrar observación de un ejemplar",
        "tags": [
          "Inventario"
        ],
        "x-implementation-status": "implemented"
      },
      "get": {
        "operationId": "InventoryController_listObservationsByCopy",
        "parameters": [
          {
            "name": "copyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Consultar observaciones de un ejemplar",
        "tags": [
          "Inventario"
        ],
        "x-implementation-status": "implemented"
      }
    },
...
    "/auth/accounts": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterAccessAccountDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Registrar cuenta de acceso",
        "tags": [
          "Autenticación"
        ],
        "x-implementation-status": "pending"
      }
    },
    "/auth/authenticate": {
      "post": {
        "operationId": "AuthController_authenticate",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthenticateAccountDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Autenticar cuenta de acceso",
        "tags": [
          "Autenticación"
        ],
        "x-implementation-status": "pending"
      }
    },
    "/auth/accounts/{accountId}": {
      "get": {
        "operationId": "AuthController_getById",
        "parameters": [
          {
            "name": "accountId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Consultar cuenta de acceso",
        "tags": [
          "Autenticación"
        ],
        "x-implementation-status": "pending"
      }
    },
...
  },
  "info": {
    "title": "CS Préstamos — API",
    "description": "Contratos HTTP declarados en el backend. El estado de implementación se indica por operación.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string",
            "maxLength": 120
          },
          "tiempoMaximoPrestamoDias": {
            "type": "number",
            "minimum": 1
          }
        },
        "required": [
          "nombre",
          "tiempoMaximoPrestamoDias"
        ]
      },
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string",
            "maxLength": 120
          },
          "tiempoMaximoPrestamoDias": {
            "type": "number",
            "minimum": 1
          }
        }
      },
      "CreateResourceDto": {
        "type": "object",
        "properties": {
          "categoriaId": {
            "type": "string",
            "format": "uuid"
          },
          "nombre": {
            "type": "string",
            "maxLength": 180
          },
          "descripcion": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "categoriaId",
          "nombre"
        ]
      },
      "UpdateResourceDto": {
        "type": "object",
        "properties": {
          "categoriaId": {
            "type": "string",
            "format": "uuid"
          },
          "nombre": {
            "type": "string",
            "maxLength": 180
          },
          "descripcion": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "CreateCopyDto": {
        "type": "object",
        "properties": {
          "codigoInventario": {
            "type": "string",
            "maxLength": 80
          }
        },
        "required": [
          "codigoInventario"
        ]
      },
      "UpdateCopyStateDto": {
        "type": "object",
        "properties": {
          "estado": {
            "type": "string",
            "enum": [
              "DISPONIBLE",
              "PRESTADO",
              "NO_DISPONIBLE"
            ]
          }
        },
        "required": [
          "estado"
        ]
      },
      "CreateCopyObservationDto": {
        "type": "object",
        "properties": {
          "descripcion": {
            "type": "string",
            "maxLength": 2000
          }
        },
        "required": [
          "descripcion"
        ]
      },
...
      "RegisterAccessAccountDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "format": "password",
            "minLength": 8
          },
          "role": {
            "type": "string",
            "enum": [
              "USUARIO",
              "OPERADOR",
              "ADMINISTRADOR"
            ]
          }
        },
        "required": [
          "userId",
          "email",
          "password"
        ]
      },
      "AuthenticateAccountDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "format": "password"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
...
    }
  }
}